bug: 46891
When the search involves a remote QO, we setup the remote search by creating a ProxiedQueryResults.  The query string for the remote search was overwriting the original query string.   This is fine if the search is successful.  But not OK if we hit  TEXT_INDEX_OUT_OF_SYNC during the search.  In that case we do the indexing, then re-search.  Since the query string was overwritten to the one for the remote search, we won't be getting any local results as requested originally.

The fix is setting the query string for the remove search on its own cloned copy of SearchParams.

http://bugzilla.zimbra.com/show_bug.cgi?id=46891