Integrate change 240698 from GNR to main
-------------
Bug: 47326 - session expiration can result in lost notifications
Make sure we don't use a notification sequence number associated with an old session. That would cause the new session to skip notifications until the sequence number caught up. One way for that to happen is on receipt of a pending NoOpResponse (whose request was made during the old session).
- cancel pending poll on refresh, send a new one
- remember stale session IDs so we can check for them
- it's really bad to inadvertently set the session ID to null, so add checks to ZmCsfeCommand.setSessionId(), and provide a new function ZmCsfeCommand.clearSessionId() to explicitly clear the session
- client command "$set:instant_notify" takes "on" as arg instead of 1
- in the debug window, show the full method name for response, eg SearchResponse

http://bugzilla.zimbra.com/show_bug.cgi?id=47326