bug: 20620
Track the number of items and unread items in the folder with the \Deleted flag set.  Also track the number of tagged unread items marked as \Deleted.  When returning these counts via SOAP, return the non-\Deleted counts, but also include the complete counts (including \Deleted) in a different attribute if the two values differ.

Make sure that IMAP uses the complete counts (i.e. including \Deleted) when doing STATUS on remote folders.

http://bugzilla.zimbra.com/show_bug.cgi?id=20620