bug: 42373
When a user deletes subfolders of a mountpoint, just *pretend* to delete them by hiding them from that client for the duration of the session.  This deals as best we can with users who try to delete mountpoints via IMAP, resulting in their IMAP client issuing DELETE requests for the remote subfolders.

Deletes of the mountpoints themselves still work (as before).  And if you really want to delete remote folders, DELETE /home/<username>/path/to/be/removed will still succeed.

http://bugzilla.zimbra.com/show_bug.cgi?id=42373