Integrate change 243622 from GNR to main
-------------
Bug: 48865 - AJAX web client loses keyboard focus
Clicks on inline elements (SPAN, NOBR) within a list view were getting handled by Dwt.ffScrollbarCheck as if they were in the scrollbar, and the subsequent twiddling of the DOM event triggered a blur in the hidden keyboard focus field. For some reason those elements have nonzero scrollHeight and scrollWidth, but zero clientHeight and clientWidth, and we thought they were DIVs with scrollbars. The fix is to make sure the values we compare are nonzero.

http://bugzilla.zimbra.com/show_bug.cgi?id=48865