bug: 10287, 6288
Initial set of server changes for the calender reminder alerts feature. This check-in would only do email reminders (no sms for now; need to do some more research on the logistics of it). For now the email reminders would be sent only to the zimbra account. We will build on top of this and soon add the ability (including the UI changes) to send reminders for an appointment to multiple email addresses. 

Here's a summary of the changes:

Added PendingNotifications as an argument to MailboxListener.mailboxChanged() interface method.
Introduced CalItemReminderService (extends MailboxListener) that listens to creation/modification/deletion of appointments and schedules reminder tasks via ScheduledTaskManager.
Introduced a (tagging) interface ScheduledTaskResult and now passing the last task result to ScheduledTaskCallback.afterTaskRun() method.
Introduced CalItemReminderTask (extends ScheduledTask) instances of which are scheduled by CalItemReminderService. The call() method of this task returns CalendarItem instance as the result.
Introduced CalItemReminderTaskCallback (implements ScheduledTaskCallback<ScheduledTaskResult>) that, after a reminder task runs, updates the next alarm on the calendar item and schedules the next reminder task as required.
Introduced a LC property "calendar_reminder_email_enabled" (default: "true") that controls whether to schedule reminder tasks.

http://bugzilla.zimbra.com/show_bug.cgi?id=10287
http://bugzilla.zimbra.com/show_bug.cgi?id=6288