bug: 49669
File the save-to-sent reply with the original message regardless of the original message's age when the original message was in a virtual conversation.

We were previously incorrectly normalizing all explicit conversation IDs under 256 to "-1", which did the wrong thing when replying to a message in a virtual conversation (whose id = -messageID).  The code still generally did the right thing anyway, since replying to a message <30 days old would go to the OPEN_CONVERSATION table and find the right message to group with anyway...

http://bugzilla.zimbra.com/show_bug.cgi?id=49669