bug: 49820
Use requested/target account as authenticated account if the authenticated account is the ZD special "local" account.

The change in ForwardAppointmentInvite and ForwardCalendarItem ensures the actual sender email address is used to set SENT-BY parameter in ORGANIZER iCalendar property and Sender MIME header.

The change in OfflineMailSender ensures the ZD local account doesn't interfere with the logic in MailSender that prevents unauthorized send-as/send-on-behalf-of.

http://bugzilla.zimbra.com/show_bug.cgi?id=49820