bug: 49887
Handle the case where we're on another server but we're trying to resolve a path that lives on our home server.  (Only case this happens is COPY after a SELECT of a remote shared folder.)  A few assumptions we were making are not valid in that case, so filled in missing logic for handling ZMailbox mountpoints.

(Wrote a version of getFolderByPathLongestMatch for ZMailbox; this should probably go in the zclient code instead of in the IMAP code.)

http://bugzilla.zimbra.com/show_bug.cgi?id=49887