Integrate change 247595 from GNR to main
-------------
Bug 49662
This bug is caused by different email-address validation strategy. At the alias-add stage, if the alias content is cleared, we think the alias name is optional, but at the stage of account-creation, the email-address can't be null. It conflicts. In order to avoid  modifying the validation strategy, I add some check when form handle with the error message. If this error message is thrown byuser-name is null at the stage of adding alias, we will clean the error status, just ignore it.

http://bugzilla.zimbra.com/show_bug.cgi?id=49662