bug:49886 Execute DB search within mailbox transaction
When ZimbraHit constructs MailItem from UnderlyingData, the UnderlyingData can
be stale because it's outside of mailbox transaction. The problem is that
MailItem generated from those stale UnderlyingData can go into cache.

This fix forces the UnderlyingData fetch and the MailItem conversion to be
executed within a same mailbox transaction. This way, UnderlyingData should
never be stale when it gets converted to MailItem.

https://reviewboard.eng.vmware.com/r/168386/

http://bugzilla.zimbra.com/show_bug.cgi?id=49886