bug: 27944
Avoid calling File.isDirectory() (which in turn calls stat(2)) on every file in blob directories.  All blob file names end in ".msg", so only call File.isDirectory() for entries that don't fit the name format.  It's possible to hand-create a directory named *.msg, but that should never happen in practice.  But the code handles that case gracefully too.

Fix the zip file reader to not choke on a zip file that contains no entry.  Such a zip file is not an empty file.  It consists solely of the end-of-central-directory record, and is 22 bytes long.  Backup can produce this "empty" zip file in the above mentioned case of intentionally malformed blob subdirectories named *.msg.

http://bugzilla.zimbra.com/show_bug.cgi?id=27944