bug: 50177 - Should use GAL contact names instead of LDAP attribute name in SearchCalendarResource
Changed ZmResource.F_mail to "email", ZmResource.F_name to "fullName" and added a new string constant:
ZmResource.F_description = "notes". 
Made ZmApptChooserTabViewPage use the correct string constants when performing a search.

We're having some difficulty integrating the change sets from bug: 42828 on our development server, and have been unable
to test this with the server-side GAL changes.
Perhaps subsequent testing of this bug should be done on a server with access to the server-side changes.
Revert this if it breaks anything.

http://bugzilla.zimbra.com/show_bug.cgi?id=50177