bug: 19222, 21875
Alternate MIME parser implementation as a state machine.  Both MimeParserInputStream and MimeParserOutputStream use the new MimeParser class and pass all bytes through it as they are read/written.  When the stream is exhausted, getMessage() returns the resulting parsed MimeMessage.

Our old parser implementation (MimePart.readContent()) is still in place for the moment.  But the new parser appears to give the same results under all tests, except for a few places where it fixes longstanding parser bugs.  Switching over to the new parser for MimeMessage constructors will come in a separate checkin.

http://bugzilla.zimbra.com/show_bug.cgi?id=19222
http://bugzilla.zimbra.com/show_bug.cgi?id=21875