bug: 50398
Correctly deal with parameters with no value.  ";" or ":" token must be pushed back to the tokenizer if encountered when expecting a parameter value, so that the delimiter token may be seen by the parameter list loop guard.  If the ":" token is not pushed back, the loop guard will consume the next token that is part of the _property_ value.  This then causes the property value to be parsed without the beginning part.  For example, ":user@domain.com" instead of "mailto:user@domain.com".

http://bugzilla.zimbra.com/show_bug.cgi?id=50398