bug: 50246
When we serialize message MIME parts to soap, there is a check to avoid outputting calendar parts as <mp> because they should be serialized as <inv> instead.  But the check only look at the main/sub content type and thus can't distinguish between a text/calendar part that contains a calendaring object and an attachment that happens to be a .ics file.

Added a further check for presence of "method" parameter in the part's Content-Type.  A true calendar part has a method and are serialized as <inv>.  Other .ics attachments are serialized as <mp>.

http://bugzilla.zimbra.com/show_bug.cgi?id=50246