bug: 25458
If a recipient resides on a remote server instead of the local server (usual case), making zimbra server act as an LMTP client/proxy to the remote server to deliver the mail.

Tested by deploying a second zimbra server instance on the same machine using the server2-reset ant target and then using zmlmtpinject command to inject messages.

http://bugzilla.zimbra.com/show_bug.cgi?id=25458