bug 39647 added migration CLI for creating Document item for each WikiItem in the user's mailbox.  the migration will leave the existing WikiItem intact.  new Document items will be created under a new folder named migrate-yyyyMMddHHmmss.  all the folder with defaultView set to Wiki will be recreated under the new migration folder. all the Documents will be recreated under the new folder. all the WikiItem will be recreated as Documents.  resulting html file will be editable using the Briefcase Document editor.
the migration preserves all the revisions if they are available.
the revision date and the user information for the revisions
are also preserved.
the migration action is repeatable as it will create a new
top level migration folder with different timestamp each time.

also added Admin SOAP API MigrateAccount to facilitate migration
within mailboxd.  no admin UI support is planned for now.

http://bugzilla.zimbra.com/show_bug.cgi?id=39647