bug: 24096
When inactivating an IMAP session (i.e. detaching it from an interactive client session but leaving it in memory catching changes), remove the session from the master reaped SessionMap.  Otherwise, they count against the quota and may cause idle interactive IMAP sessions to be dropped when the IMAP session count limit is exceeded.

http://bugzilla.zimbra.com/show_bug.cgi?id=24096