bug: 39856 - remove digest/size calculation from ParsedMessage
Removed size and digest calculation from ParsedMessage.  This
required me to remove the size hint from the
StoreManager.storeIncoming() API, since there are times when
we create a MimeMessage in memory and don't know the size
until the data is written to disk.  BlobBuilder now reads the beginning
of the stream into a buffer, in order to determine whether the
resulting blob will be compressed or not.

There were a few APIs that pass around a ParsedMessage
and depend on the message size.  I modified these APIs to
pass the size separately.

http://bugzilla.zimbra.com/show_bug.cgi?id=39856