bug: 51242
Introduced draft flag for Invite objects.  This is not to be confused with the draft flag at item level.  Invite-level draft flag should be used by organizer only, and it should be set on invites that have changes that haven't been sent to attendees.  Setting it is completely at the discretion of the calendar client app and server doesn't automatically set it.  It cannot be set on cancel invites (server filters it out) because cancelled instances cannot be edited to clear the draft flag.

Clients should set invite-level draft flag by adding draft="1" in <comp> element in CreateAppointmentRequest, ModifyAppointmentRequest, CreateAppointmentExceptionRequest, SetAppointmentRequest, etc.

Various get calls like GetAppointmentResponse and GetMsgResponse return the draft flag in the same format, as <comp draft="1">.  Client should use this for a visual queue.

Attributes of item-level draft flag is changed to (MESSAGE_ONLY | CALITEM_ONLY).  Previously the flag was for email messages only.  An appointment item's item-level draft flag is set if appointment has at least one draft invites.  Setting the item-leve draft flag allows searching by "is:draft" query.

The Flag.FLAG_GENERIC constant has been redefined to a non-zero value.  The flag attributes are not persisted/serialized, so it's okay to redefine the values.  The FLAG_GENERIC value had to be changed to a non-zero value to allow OR'ing of multiple attribute bits.  (Draft flag's attributes have both FLAG_MESSAGE_ONLY and FLAG_CALITEM_ONLY bits set.)

All of the above changes apply to tasks as well as appointments.

http://bugzilla.zimbra.com/show_bug.cgi?id=51242