bug 51672
Nginx uses a custom escape/unescape schema for auth_http. It simply encodes " " to "%20" and "%" to "%25". This is not a standard and thus no existing Java utility can decode this. Therefore implement an Java edition of unescape in NginxLookupExtension which exactly matchs the approach used by nginx.

Because nginx only escapes the username and password of route lookup http request, but never unescape the response, only unescape in NginxLookupExtension is necessary.

http://bugzilla.zimbra.com/show_bug.cgi?id=51672