Bug: none
Selenium Harness:  Mobile, HTML, and Admin harnesses aren't using the correct URL in STAF

Setting the appType is occuring in ExecuteHarnessMain.parseArgs().  It should probably happen in the app-specific code, such as CommonTest.

When running in STAF, the ExecuteHarnessMain methods main() and parseArgs() never happens.  So, any logic in those methods isn't happening for STAF. 

Updated the code so the appType is set (again) in the app-specific code.