bug: 52689
Removed "assert(sessionId == null)" at the beginning of WaitSetAccount.createSession().  sessionId can be non-null on entry to createSession if the associated Mailbox is not loaded in memory or the session identified by the session id is no longer around.  Basically it's the condition for WaitSetAccount.getSession() to return null when sessionId member field is non-null.

http://bugzilla.zimbra.com/show_bug.cgi?id=52689