Bug 52698
IE sometimes returns "auto" for top and left attributes. This was causing a NaN value and I check for it and change it to 0 in this case. It seems to work.
I also take this opportunity to check in a little fix for AjxUtil.assert stack-trace generation for IE. this code is not really used but I used it to my advantage after fixing, to generate a call stack, in order to solve this bug.

http://bugzilla.zimbra.com/show_bug.cgi?id=52698