Bug 50165
fixed API usage bug (needs to create range from the selection before using the text) and also had to call our focus() since focus was lost. In addition I
fixed the case where the user highlights html (e.g. bold items, etc) and
removes the text field - it inserts the html to the innerHTML of the link. All
this is IE only.

http://bugzilla.zimbra.com/show_bug.cgi?id=50165