bug: 16343
Re-organised client WSDL tests which require Zimbra Server to be setup and
configured so that they are only run from "ant wsdl-test"
"ant test" will run tests that do not require the server to be installed.

Considered mirroring ZimbraServer/src/java/com/zimbra/qa/unittest framework
but those classes end up in the deployed jar which is not desirable in this
case.

http://bugzilla.zimbra.com/show_bug.cgi?id=16343