bug: 52914
Added limit and offset attrs in SearchCalenderResourceRequest, added paginationSupported="1|0" in SearchCalenderResourceResponse.

For LDAP search, pagination is not supported, paginationSupported will be set to 0 in the response.  It indicates limit and offset was not honored, and client should not try to page through it.   Numebr of entries returned is capped by the LC key calendar_resource_ldap_search_maxsize, default to 1000.

http://bugzilla.zimbra.com/show_bug.cgi?id=52914