Integrate change 259475 from HELIX to main
-------------
Bug 51852
In webkit-based SF and Chrome, the activated event of enter key is 
"onkeypress", however the handler codes are created for incorrect 
event "onkeydown" in SF and Chrome.
Comparing with FF/IE, the html code for enterkey can be created
correctly. For IE, "onkeydown" is used; and for FF, the AjxEnv.isNav 
helps it create the "onkeypress". Since the SF/Chrome will be activated
by the "onkeypress" event, so the bug can be fixed by catching the
"onkeypress" event but not by "onkeydown" event. In the bugfixing,
the html creator for event handler has been overrided to create
corresponding "onkeypress" handling codes for _TEXTAREA_ component. 
Reviewed:
https://reviewboard.eng.vmware.com/r/189872/
The bugfixing has been tested in SF5/Chrome7/FF3.6/IE8.

http://bugzilla.zimbra.com/show_bug.cgi?id=51852