Integrate change 259522
bug: 52276 - Place toast messages in area specified by skin

Introduced new optional element with id "skin_container_toast".
When this element is present in a skin, toasts will be placed relative to that element rather than to the zshell.
The "location" parameter still specifies where in the element the toast goes, so a location of "C" places the toast in the middle of the element, "W" along the left edge of the element, etc...

http://bugzilla.zimbra.com/show_bug.cgi?id=52276