Bug 53551
resize message body on autocomplete in bubbles mode. It was not resizing on all browsers, but on IE it cause a more visible implication as when the user focuses on the message body textarea, it was scrolling it up to allow for all of it to be visible, thus top of container got out of view.

http://bugzilla.zimbra.com/show_bug.cgi?id=53551