bug: 38834 - Improved console handling for CLI utilities
Turned on JLine command line editing when zmprov
and zmmailbox are running in interactive mode.

Ideally we should use rlfe so we support standard key bindings
on Linux.  This will require updating the zmprov and zmmailbox
launch scripts to do duplicate command line parsing, in order
to determine whether the command is running in interactive
mode.

http://bugzilla.zimbra.com/show_bug.cgi?id=38834