bug: 54072
The assertion was happening because the start and end times of the requested time range were equal.  Such a free/busy request is meaningless.  Assertion is avoided by returning an INVALID_REQUEST exception up front: "End time must be after Start time".  This exception existed all along, but contrary to the text it was only checking if end < start.  Now the exception is thrown if end <= start.

http://bugzilla.zimbra.com/show_bug.cgi?id=54072