bug: 19222
Setting a body part's content now correctly sets the Content-Transfer-Encoding header.  Likewise, parsing the content of a body part via the MimeParserInputStream will choose a transfer encoding and set the header appropriately.

When adding/setting headers, "Received" and "Return-Path" are prepended to the block rather than appended.

Don't allow the parser's endParse() method to be executed more than once.  Throw an exception if we try to parse more data after the endParse().

Setting a part's filename always updates both its Content-Type and Content-Disposition headers.

Simplified the the base64 encoder's logic generating the terminal CRLF.

When overriding the content of a JavaMailMimeBodyPart, *always* tranfer the old content's headers to the new content regardless of whether or not there's already a Multipart parent associated with the part.

Added JavaMailMimeMessage.getMessageStream() to return a stream to the entire message (including headers).  In the case that we're not using the Zimbra parser, we use the stream from an appropriate DataHandler, which goes via a PipedInputStream and a separate thread.

Consolidated the shim DataSource parsing code into JavaMailMimeBodyPart.parsePart().

Caught a few shim class method that weren't actually shimmed to the superclass.

Set the Content-Type correctly on uudecoded files' data sources.

Skip the transfer encoding header repair step if we're using the Zimbra parser.

http://bugzilla.zimbra.com/show_bug.cgi?id=19222