Integrate change 261963 from HELIX to main
-------------
Bug: 54229 - Improve troubleshooting logging
Add some logging functionality to AjxDebug that can be called statically, so that it's available in prod mode. It's intended for high-priority troubleshooting, so should be used sparingly.
To troubleshoot a particular problem, define a type for it and log messages using that type. A queue of 200 messages will be kept in memory. To see them, run a search:
$set:log [type]
A dialog will pop up containing the messages. To change the number of messages kept, add a number:
$set:log [type] [size]
To start with, we support types "rpc" and "notify".
- got rid of the dummy DBG object we create in prod mode; it's much cleaner just to load the two small files and instantiate AjxDebug

http://bugzilla.zimbra.com/show_bug.cgi?id=54229