Integrate change 262088 from GNR to main
-------------
Bug 45129
This bug is caused by dismatch between mousedown and mouseup event on the same element. When we do fast click, sometimes, the next mouseup event is not on the same element with the previous mousedown event. 
It is easy for us for reproduce this bug as follows. Mousedown on one TreeItem then move the mouse, mouseup on the other TreeItems.
       In fact, we doesn't need to do focus on the TreeItem mousedown event for in the mouseUp event handle function, we has do focus. And only mousedown and mouseup happens on the same element can treat as a click event triggers. 

http://bugzilla.zimbra.com/show_bug.cgi?id=45129