Integrate change 263550 from HELIX to main
-------------
Bug 54581
Remove the unused attributes of data source polling as recommended
by Brian. Also, change the description of "Minimum polling interval".
The change (NO. 263310) is for HELIX and main branch because GNR is 
not containing so much attributes which is asked for remove. The other 
change (NO.263325) will focus on the modification on GNR, in which 
zimbraDataSourcePollingInterval will be deleted as in HELIX and 
zimbraDataSourceMinPollingInterval will be added (for Charles' 
suggestion). Change separation is for the convenience of code 
integration.
Reviewed:
https://reviewboard.eng.vmware.com/r/194971/

http://bugzilla.zimbra.com/show_bug.cgi?id=54581