Bug 54409
load the overview controller on demand for child windows. It's needed for the mail message attach dialog.  This required removing change listeners on unload of the child window, to prevent exceptions due to listener from dead window still being registered. Fixed a bug in ZmNewwindow.unload - wrong attribiute was used for the command. It's window.newWindowCommand.

http://bugzilla.zimbra.com/show_bug.cgi?id=54409