bug: 53182
Switch the ThreadPoolExecutor's work queue to be a SynchronousQueue.  As a result, the thread pool will expand beyond the core pool size up to the maximum thread count as needed to support new connections.  When the pool is full, incoming connections are accepted, an ERROR is logged to mailbox.log, a protocol-appropriate error message is optionally written and flushed to the client, and the socket is closed.

http://bugzilla.zimbra.com/show_bug.cgi?id=53182