Integrate change 263265 from HELIX to main
-------------
Bug 53579
The status of account type is controlled by calculating the available 
COS limit and the condition whether or not adopted COS is same as
current type choice. This case is correct during open existing
account view. However, it is not wholly suitable in account creation
because creation would ask for one from current COS quota (not apply
existing or allocated account type like opening a existing account). If
default COS name for new account is same as specific COS name which is
over account limit, it will be marked as "checked". This is reason 
incurred the bug in NewAccountWizard. Therefore, the bugfixing 
differentiates both conditions and control the status of account type
currently.
The bugfixing has been tested in FF3.6/Chrome7. 
Reviewed:
https://reviewboard.eng.vmware.com/r/194627/

http://bugzilla.zimbra.com/show_bug.cgi?id=53579