Integrate change 267495 from HELIX to main
-------------
Bug: 54903 - Autocomplete list stays up
Create a general solution for detecting mouse clicks outside of a given control or element. The caller provides a control or element, and a listener to call when the user clicks outside that control or element.
The clients so far are: right-click menus, autocomplete lists, and undo toast.
Blurring the window is also covered, mostly. IE has a broken implementation of window.onblur, and using document.onfocusout can be tricky, so window blur is not covered for IE.
Also:
- move the input to after the last bubble
- dynamically size the input to be 120 pixels wide, or the width of user input plus 30, whichever is greater
- fix use of the Delete key
- handle expansion of DL with 0 or 1 member(s)
- fix problems with expanding a DL when one has already been expanded
- create global key event listeners in DwtKeyboardMgr
Testing
-------
While a right-click menu, autocomplete list, or undo toast is up, click anywhere outside the menu, list, or toast. The menu, list, or toast should go away, and the click should have the same same result it would have had without a menu, list, or toast being up. Areas to click include:
- a folder in the overview
- the search bar's input
- blank area at the bottom of the overview, or any blank area
- user name
- list view
- msg view, esp. an HTML msg
- compose view in HTML mode, in the body
- in the browser's location bar (except IE)
- on another window (except IE)
Make sure that right-click menus, autocomplete, and undo work as expected.

http://bugzilla.zimbra.com/show_bug.cgi?id=54903