Bug 52983
The cause of the bug is the following sequence of events:
1. "s" keyDown event triggering ZmDialog.prototype.popup()
2. ZmDialog.prototype.popup setting focus on the input field.
3. "s" keyUp event triggering ZmChooseFolderDialog.prototype._handleKeyUp. - this does the search which erases the previous result highlighted.
The keyUp is called on the input field, but does not affect its value since the keyDown was not done there.
The confusing thing was when we break in firebug before the focus on the field, the problem did not happen - this is since the keyUp event had time to be called before the focus thus NOT triggering the
ZmChooseFolderDialog.prototype._handleKeyUp.
I eventually solved it by adding a keyDown handler to the input field and
setting a flag there indicating to _handleKeyUp that the key it processes is legit and not a result of the "s" shortcut.
It might be better to not call the keyAction shortcut handlers on keyDown but rather only on keyUp, but it might be complicated since the code is not simple and handles combination such as CTRL+key where keyDown is required.  So for now I think my solution is simple and safe.
I also fixed the search so it chooses the FIRST folder from the active account that matches, instead of last one like before. (that caused the last one to be the default in the first time, which was not that great).

http://bugzilla.zimbra.com/show_bug.cgi?id=52983