Integrate change 267713 from HELIX to main
-------------
Bug 55020
The domain name is not case-sensitive in the server-side. For example, when we get domain info and set domain name to "Zimbra-Mail-Server", the response in domain info's name value is zimbra-mail-server. We use name value in domain-info as a key in hash-table. But in js, the key is case-sensitive, so the cache is always missed in this case.
And update the method for get old domain-name, the EmailAddr_XFormItem's oldDomain semantics is a little different from our requirement. The EmailAdd_XFormItem's oldDomain is only updated when we input a new Domain. For example, if the old one @old, and the old-domain will be old. After we change the domain to A@domain, the old part will be old. But we change it to B@domain. It will be still old, but for our requirement it should be domain. So just use the full-name is stored in XForm to get the update the old domain-name.  

http://bugzilla.zimbra.com/show_bug.cgi?id=55020