Bug 54975
I reverted my fix to bug 51962 which dealt with the issue with the header not getting a scroll bar even when it was too big.
I also fixed the calculation of the height of the iframe since doc.body.scrollHeight was giving the height available to the iframe, or
something like that, but all we care about is the height of the content and iframe.contentWindow.document.documentElement.scrollHeight seems to be good.

http://bugzilla.zimbra.com/show_bug.cgi?id=54975