Integrate change 268246 from HELIX to main
-------------
Bug 55020
When this account has limited right, its getattrs and setattrs are null. In this way, the domain obj's id and name are not initialized. But in the domain's cache, it will use null value as a key in hash table. It will make the cache always missed. We do init when we get right from server for the reponse also has name and id info. For all the item has this issue, just init for all the ZaItem.
      Also, and a check for domain name and id, for we don't want add null value as a key in hashtable.

http://bugzilla.zimbra.com/show_bug.cgi?id=55020