bug: 55236
When Message.addMessage() is called with noICal==true, deactivate any calendar-related auto-reply/forward emails.  The calendar attachment is still processed to store parsed data in the message's metadata, so the info can be displayed when message is viewed.

Propagate the noICal argument through the filtering API, so the email delivery can request full calendar processing and POP/IMAP import can suppress it.

http://bugzilla.zimbra.com/show_bug.cgi?id=55236