bug: 7049
Relax the requirement that the REST request must come in on the amdin port for admin accessing a non-active account.  For admin console -> view mail, the web client issues a REST request to get all contacts of the account.  The REST comes in on the domain public URL, which we cannot force it to be on the admin port.    For accessing non-active accounts in REST, we already check that the auth token must be either an admin token or a delegated token, both can only be acquired on the admin port earlier.

http://bugzilla.zimbra.com/show_bug.cgi?id=7049