Integrate change 269236 from HELIX to main
-------------
Bug 55263
This bug is caused by changelist 254750 for bug 39679.
For ZaGlobalStatsView use the wrong pageView to generate the TabView, it leads to the ZaGlobalStatsView and ZaServerStatsView using the same id to represent two div. It will make the IE feel confuse to update the div's content by id.
I think the right step to reproduce bug is:
1. Log into the Admin-Console.
2. Click on the  Server Statistics. 
3. Click on the any tab(Message Count/Message Volume/Anti-Spam).
4. Click on one server's name. 
5. Click on the tab of this server which click on the step 3.

http://bugzilla.zimbra.com/show_bug.cgi?id=55263