bug: 49199, 55186
Indicate soap trace of a resumed request as "(resumed)".  Without it, the initial trace and resumed trace look identical and the reader can be easily fooled into thinking the client made two requests.  It's even more confusing because there is only one response trace logged.

Log a message to the trace log when WaitSetRequest and NoOpRequest puts the request into suspension.  This message will serve as the match for the initial request trace.

http://bugzilla.zimbra.com/show_bug.cgi?id=49199
http://bugzilla.zimbra.com/show_bug.cgi?id=55186