Integrate change 270744 from HELIX to main
-------------
Bug 51763
the problem was caused by part of the fix to bug  55277.
This bug was caused by newlines ("\n"s) converted to <br>s.
According to Lars' fix it was for converting to HTML, so I changed it so it only does this conversion when switching from text to html compose. The line breaks are still added to the signature (and actually in general to the message itself on IE) but I didn't know what to do in order not to break what Lars was trying to fix. This fix at least fixes the common case, of composing in html (as default) without switching compose mode back and forth.

http://bugzilla.zimbra.com/show_bug.cgi?id=51763