Integrate change 270977 from zimbra to main
-------------
Bug 55660
The bug is occurred from the un-deletion of temporary added member list.
The bugfixing adds the checking when accounts are required to remove from
member list. If some accounts are just temporary added member (not existed
accounts), the fixing will get rid of this kinds of accounts.
The bugfxing has been tested in FF3.6/Chrome8.
Reviewed:
https://reviewboard.eng.vmware.com/r/201408/

http://bugzilla.zimbra.com/show_bug.cgi?id=55660