Integrate change 271176 from HELIX to main
-------------
Bug 50435
the original checkbox class name might not have been set already for a selected item (it was set if the checkbox was clicked at least once or the tooltip was active, but not otherwise). So we know if it is undefined, the checkbox is not checked. Previously it was not doing anything it if was not set.

http://bugzilla.zimbra.com/show_bug.cgi?id=50435