bug: 50991, 28046
When sending notifications to Mailbox listeners, notify with a point-in-time snapshot of the changeset rather than with pointers to live items.  This way, subsequent changes to the mailbox will not mutate the changeset, allowing us to safely cache it.

Tweak Folder.asChild() so that callers can indicate whether or not the added child is due to constructing an existing folder tree or due to actually adding a new subfolder.  (In the former case, we don't want to mark the parent folder as modified...)

Insert an extra check during ImapHandler.fetch() to catch messages not marked as expunged but which no longer exist.  (This is an error.)  Instead, silently expunge the message and return a stub; subsequent operations will clear the bad entry from the ImapFolder object during collapseExpunged().

http://bugzilla.zimbra.com/show_bug.cgi?id=50991
http://bugzilla.zimbra.com/show_bug.cgi?id=28046