Integrate change 272027 from HELIX to zimbra
-------------
Bug 52541
A general way to solve the text cut on the button in the admin-console. If the text is too long, make the text short and add three '.'. If there is no tooltip set, the orginal text will be set as the tooltip. 
Also, modify the direct reference to appCtxt in AjxStringUtil.js to avoid js error.

http://bugzilla.zimbra.com/show_bug.cgi?id=52541