bug: 18288,54360
Added a couple of new filter actions - reply and notify.

Soap:

  <actionReply>
    <content>{body-template}</content>
  </actionReply>

  <actionNotify a="email-address" [su="subject-template"] [maxBodySize="{num-bytes}"]>
    <content>{body-template}</content> ?
  </actionReply>

  body-template/subject-template: Can contain variables such as ${SUBJECT}, ${TO}, ${CC}, etc (basically ${any-header-name}; case not important), plus ${BODY} (text body of the message).

Sieve:

  reply "body-template"
  notify "email-address" "subject-template" "body-template" [max-body-bytes]

body-template text is always encoded in the sieve script as a multi-line form starting with keyword "text:" (see section 2.4.2 of RFC 3028).

zmmailbox:

  reply "body-template"
  notify "address" "subject-template" "body-template" ["max-body-size(bytes)"]

Added unit tests. Introduced <trueTest> for the simple "true" filter test/condition defined in section 5.10 of the RFC. This helps in writing unit tests for testing filter actions when you want the filter condition to always evaluate to true.

http://bugzilla.zimbra.com/show_bug.cgi?id=18288
http://bugzilla.zimbra.com/show_bug.cgi?id=54360