Integrate change 273414 from HELIX to zimbra
-------------
Bug: 55777 - Navigate bubbles with arrow keys
Allow the user to change which bubble is selected by using the left and right arrow keys. Going right from the last bubble will select the INPUT. Going left from an empty INPUT will select the last bubble.
Since we really really don't want to mess with focus, and focus is normally required for triggering shortcuts, we override the keyboard mgr's retrieval of the focus object by returning it as an expando property on the keydown event via the callback we registered. That way, the override is only present for the handling of that specific key event.

http://bugzilla.zimbra.com/show_bug.cgi?id=55777