Integrate change 273535 from HELIX to zimbra
-------------
Bug 40184
implemented a ZmProgressController that accepts folder list, and a work
specification. It runs the search to get the messages ids of the messages in
the folders, in chunks, and calls the work object to do it's job. The work
object in this case is defined by ZmFilterRulesController.ZmFilterWork - it
simply runs the filters on the passed message ids (and keeps track of
progress).

http://bugzilla.zimbra.com/show_bug.cgi?id=40184