Integrate change 273949 from HELIX to zimbra
-------------
Bug: 55460 - Rubber-band selection of address bubbles
Allow the user to click and drag a rectangle to select bubbles. Also, fix shortcuts in Chrome.
DwtDragBox models the rectangle, which is drawn as a dotted outline. It is similar to but separate from DnD support. I also cleaned up DnD support in DwtControl to make it easier to follow. A control that supports a drag box will start drawing it on click and drag, and remove it on mouseup. A listener within the control will be notified as changes occur.
The address input checks to see which bubbles are overlapped by the box, and selects them.

http://bugzilla.zimbra.com/show_bug.cgi?id=55460