Integrate change 274088 from HELIX to zimbra
-------------
Bug: 55792 - After cut of bubble, left with space that you can't delete
In edit mode, INPUT has a pad of 10 (rather than 30 while not in edit mode), so that there's not a lot of space between the INPUT and the next bubble. The INPUT will have a minimum width of 30 in edit mode. Normally it will size to the width of the content plus the pad of 10.
- also fix problem with context menu for an address that has no friendly part

http://bugzilla.zimbra.com/show_bug.cgi?id=55792